backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
PyJWT==2.9.0
web3==6.19.0
